+++ src/main/index.ts
@@ -1,3 +1,3 @@
 export * from './model.js';
-export * from './widget.js';
+export * from './view-model.js';
 export { App } from './app.js';